+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/Navigation.kt
@@ -19,6 +19,7 @@ import androidx.navigation.navArgument
 import cash.z.ecc.android.sdk.Synchronizer
 import cash.z.ecc.android.sdk.model.ZecSend
 import co.electriccoin.zcash.spackle.Twig
+import co.electriccoin.zcash.ui.NavigationArgs.ADDRESS_TYPE
 import co.electriccoin.zcash.ui.NavigationArgs.UPDATE_CONTACT_ID
 import co.electriccoin.zcash.ui.NavigationArguments.MULTIPLE_SUBMISSION_CLEAR_FORM
 import co.electriccoin.zcash.ui.NavigationArguments.SEND_CONFIRM_AMOUNT
@@ -37,6 +38,7 @@ import co.electriccoin.zcash.ui.NavigationTargets.EXCHANGE_RATE_OPT_IN
 import co.electriccoin.zcash.ui.NavigationTargets.EXPORT_PRIVATE_DATA
 import co.electriccoin.zcash.ui.NavigationTargets.HOME
 import co.electriccoin.zcash.ui.NavigationTargets.NOT_ENOUGH_SPACE
+import co.electriccoin.zcash.ui.NavigationTargets.QR_CODE
 import co.electriccoin.zcash.ui.NavigationTargets.SCAN
 import co.electriccoin.zcash.ui.NavigationTargets.SEED_RECOVERY
 import co.electriccoin.zcash.ui.NavigationTargets.SEND_CONFIRMATION
@@ -67,6 +69,8 @@ import co.electriccoin.zcash.ui.screen.exchangerate.optin.AndroidExchangeRateOpt
 import co.electriccoin.zcash.ui.screen.exchangerate.settings.AndroidSettingsExchangeRateOptIn
 import co.electriccoin.zcash.ui.screen.exportdata.WrapExportPrivateData
 import co.electriccoin.zcash.ui.screen.home.WrapHome
+import co.electriccoin.zcash.ui.screen.qrcode.WrapQrCode
+import co.electriccoin.zcash.ui.screen.receive.model.ReceiveAddressType
 import co.electriccoin.zcash.ui.screen.scan.WrapScanValidator
 import co.electriccoin.zcash.ui.screen.seedrecovery.WrapSeedRecovery
 import co.electriccoin.zcash.ui.screen.send.ext.toSerializableAddress
@@ -285,6 +289,13 @@ internal fun MainActivity.Navigation() {
             val contactId = backStackEntry.arguments?.getString(UPDATE_CONTACT_ID).orEmpty()
             WrapUpdateContact(contactId)
         }
+        composable(
+            route = "$QR_CODE/{$ADDRESS_TYPE}",
+            arguments = listOf(navArgument(ADDRESS_TYPE) { type = NavType.IntType })
+        ) { backStackEntry ->
+            val addressType = backStackEntry.arguments?.getInt(ADDRESS_TYPE) ?: ReceiveAddressType.Unified.ordinal
+            WrapQrCode(addressType)
+        }
     }
 }
 
@@ -464,6 +475,7 @@ object NavigationTargets {
     const val HOME = "home"
     const val CHOOSE_SERVER = "choose_server"
     const val NOT_ENOUGH_SPACE = "not_enough_space"
+    const val QR_CODE = "qr_code"
     const val SCAN = "scan"
     const val SEED_RECOVERY = "seed_recovery"
     const val SEND_CONFIRMATION = "send_confirmation"
@@ -478,4 +490,5 @@ object NavigationTargets {
 
 object NavigationArgs {
     const val UPDATE_CONTACT_ID = "contactId"
+    const val ADDRESS_TYPE = "addressType"
 }