+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/qrcode/util/QrCodeImageGenerator.kt
@@ -1,4 +1,4 @@
-package co.electriccoin.zcash.ui.screen.receive.util
+package co.electriccoin.zcash.ui.screen.qrcode.util
 
 import androidx.compose.ui.graphics.ImageBitmap